Restructure App.tsx and VS Dashboard
