+++ frontend/src/components/index.ts
@@ -9,7 +9,8 @@ export * from './SelectFilter';
 export * from './AuthForm';
 export * from './FileInput';
 export * from './ImportExport';
-export * from './RouteGuard';
+export * from './Routes/RouteGuard';
+export * from './Routes/Routes';
 export * from './FilterDrawer/SearchBar';
 export * from './TaggedInput';
 export * from '../pages/Domains/DomainDetails';
