+++ frontend/src/pages/VulnerabilityScanDash/VulnerabilityScan.tsx
@@ -1,22 +1,15 @@
-import React, { useMemo } from 'react';
+import React from 'react';
 import { withSearch } from '@elastic/react-search-ui';
 import { Box, BoxProps, Grid, CircularProgress, Link } from '@mui/material';
 // Context & hooks
 import { ContextType, useAuthContext } from 'context';
 import { useVulnScanData } from 'hooks/useVulnScanData';
-import { useCheckUserState } from 'hooks/useCheckUserState';
 import { useMaintenanceNotifications } from '@/hooks/useMaintenanceNotifications';
 import { useClearFiltersOnMount } from 'hooks/useClearFiltersOnMount';
 import { useOrgInfo } from 'hooks/useOrgInfo';
-// Constants & types
-import { ORGANIZATION_FILTER_KEY } from 'components/FilterDrawer/VSDashRegionAndOrgFilters';
-import { OrganizationShallow } from 'components/FilterDrawer/RegionAndOrganizationFilters';
 // Shared components
-import { LoginBlockedDialog } from 'components/LoginBlockedDialog';
-import { UpdateStateForm } from 'components/Register';
 import InfoLabel from 'components/Dashboard/InfoLabel';
 import NoDataMessage from 'components/Dashboard/NoDataMessage';
-import InvitePendingCard from 'components/Dialog/InvitePendingCard';
 import PageSection from 'components/Dashboard/PageSection';
 // Local dashboard components
 import KeyMetrics from './KeyMetrics';
@@ -61,25 +54,9 @@ const keyMetricsStyle = {
 
 export const VulnerabilityScan: React.FC<ContextType> = (props) => {
   const { filters, removeFilter } = props;
-  const {
-    currentOrganization,
-    user,
-    apiGet,
-    userMustSign,
-    logout,
-    isLoggingOut
-  } = useAuthContext();
+  const { currentOrganization, user, apiGet, userMustSign } = useAuthContext();
 
-  const { isUpdateStateFormOpen, setIsUpdateStateFormOpen } = useCheckUserState(
-    user,
-    isLoggingOut
-  );
-  const {
-    isLoginBlockedDialogOpen,
-    setIsLoginBlockedDialogOpen,
-    maintenanceNotification,
-    setMaintenanceNotification
-  } = useMaintenanceNotifications(user, apiGet, userMustSign);
+  const {} = useMaintenanceNotifications(user, apiGet, userMustSign);
 
   useClearFiltersOnMount(filters, removeFilter);
 
@@ -91,47 +68,6 @@ export const VulnerabilityScan: React.FC<ContextType> = (props) => {
     error
   } = useVulnScanData(orgId ? orgId : currentOrganization?.id);
 
-  if (isUpdateStateFormOpen) {
-    return (
-      <UpdateStateForm
-        open={isUpdateStateFormOpen}
-        user_id={user?.id ?? ''}
-        onClose={async () => {
-          setIsUpdateStateFormOpen(false);
-          const updatedUser = await apiGet('/users/me');
-          if (updatedUser?.state && user?.user_type !== 'globalAdmin') {
-            const notifications = await apiGet('/notifications');
-            const active = notifications.find(
-              (n: any) =>
-                n.status === 'active' &&
-                n.maintenance_type === 'major' &&
-                new Date(n.start_datetime) <= new Date() &&
-                new Date(n.end_datetime) >= new Date()
-            );
-            if (active && updatedUser.user_type !== 'globalAdmin') {
-              setMaintenanceNotification(active);
-              setIsLoginBlockedDialogOpen(true);
-            }
-          }
-        }}
-      />
-    );
-  }
-  if (isLoginBlockedDialogOpen && maintenanceNotification) {
-    return (
-      <LoginBlockedDialog
-        open={isLoginBlockedDialogOpen}
-        message={maintenanceNotification.message}
-        onClose={() => {
-          setIsLoginBlockedDialogOpen(false);
-          logout();
-        }}
-      />
-    );
-  }
-  if (user?.invite_pending) {
-    return <InvitePendingCard />;
-  }
   if (loading) {
     return (
       <PageSection>
